% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{allnames}
\alias{allnames}
\title{allnames}
\format{
data.frame: 98012  obs. of 134 variables
}
\source{
\url{https://github.com/avehtari/ROS-Examples/tree/master/Names/data}
}
\usage{
allnames
}
\description{
allnames data set for Names example in Regression and Other Stories.
}
\keyword{datasets}
