% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{beauty}
\alias{beauty}
\title{beauty}
\format{
data.frame: 463 obs. of 8 variables
}
\source{
\url{https://github.com/avehtari/ROS-Examples/tree/master/Beauty/data/}
}
\usage{
beauty
}
\description{
Beauty data set for Beauty example in Regression and Other Stories.
}
\keyword{datasets}
