% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{births}
\alias{births}
\title{births}
\format{
data.frame: 79 obs. of 3 variables
}
\source{
\url{https://github.com/avehtari/ROS-Examples/tree/master/AgePeriodCohort/data/}
}
\usage{
births
}
\description{
Births data set for AgePeriodCohort example in Regression and Other Stories.
}
\keyword{datasets}
