% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bypass}
\alias{bypass}
\title{bypass}
\format{
data.frame: 200 obs. of 4 variables
}
\source{
\url{https://github.com/avehtari/ROS-Examples/tree/master/Bypass/}
}
\usage{
bypass
}
\description{
Bypass data set for Bypass example in Regression and Other Stories.
}
\keyword{datasets}
