% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cc2}
\alias{cc2}
\title{cc2}
\format{
data.frame: 4381 obs. of 64 variables
}
\source{
\url{https://github.com/avehtari/ROS-Examples/tree/master/Childcare/data/}
}
\usage{
cc2
}
\description{
cc2 data set for Childcare example in Regression and Other Stories.
}
\keyword{datasets}
