% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{chile}
\alias{chile}
\title{chile}
\format{
data.frame: 2591 obs. of 11 variables
}
\source{
\url{https://github.com/avehtari/ROS-Examples/tree/master/ChileSchools/data/}
}
\usage{
chile
}
\description{
Chile data set for ChileSchools example in Regression and Other Stories.
}
\keyword{datasets}
