% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{congress}
\alias{congress}
\title{congress}
\format{
data.frame: 435 obs. of 9 variables
}
\source{
\url{https://github.com/avehtari/ROS-Examples/tree/master/Congress/data}
}
\usage{
congress
}
\description{
Congress data set for Congress example in Regression and Other Stories.
}
\keyword{datasets}
