% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{earnings}
\alias{earnings}
\title{earnings}
\format{
data.frame: 1816 obs. of 15 variables
}
\source{
\url{https://github.com/avehtari/ROS-Examples/tree/master/Earnings/data}
}
\usage{
earnings
}
\description{
Earnings data set for Earnings example in Regression and Other Stories.
}
\keyword{datasets}
