% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{electric}
\alias{electric}
\title{electric}
\format{
data.frame: 192 obs. of 7 variables
}
\source{
\url{https://github.com/avehtari/ROS-Examples/tree/master/ElectricCompany/data}
}
\usage{
electric
}
\description{
Electric data set for ElectricCompany example in Regression and Other Stories.
}
\keyword{datasets}
