% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{golf}
\alias{golf}
\title{golf}
\format{
data.frame: 19 obs. of 3 variables
}
\source{
\url{https://github.com/avehtari/ROS-Examples/tree/master/Golf/data}
}
\usage{
golf
}
\description{
Golf data set for Golf example in Regression and Other Stories.
}
\keyword{datasets}
