% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{heads}
\alias{heads}
\title{heads}
\format{
data.frame: 11 obs. of 2 variables
}
\source{
\url{https://github.com/avehtari/ROS-Examples/tree/master/HeadsSchools/data/}
}
\usage{
heads
}
\description{
Heads data set for Coins example in Regression and Other Stories.
}
\keyword{datasets}
