% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{health}
\alias{health}
\title{health}
\format{
data.frame: 30 obs. of 3 variables
}
\source{
\url{https://github.com/avehtari/ROS-Examples/tree/master/HealthExpenditure/data}
}
\usage{
health
}
\description{
Health data set for HealthExpenditure example in Regression and Other Stories.
}
\keyword{datasets}
