% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{heights}
\alias{heights}
\title{heights}
\format{
data.frame: 5524 obs. of 2 variables
}
\source{
\url{https://github.com/avehtari/ROS-Examples/tree/master/PearsonLee/data}
}
\usage{
heights
}
\description{
Heights data set for PearsonLee example in Regression and Other Stories.
}
\keyword{datasets}
