% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{helicopters}
\alias{helicopters}
\title{helicopters}
\format{
data.frame: 20 obs. of 4 variables
}
\source{
\url{https://github.com/avehtari/ROS-Examples/tree/master/Helicopters/data}
}
\usage{
helicopters
}
\description{
Helicopters data set for Helicopters example in Regression and Other Stories.
}
\keyword{datasets}
