% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hibbs}
\alias{hibbs}
\title{hibbs}
\format{
data.frame: 16 obs. of 5 variables
}
\source{
\url{https://github.com/avehtari/ROS-Examples/tree/master/ElectionsEconomy/data}
}
\usage{
hibbs
}
\description{
Hibbs data set for ElectionsEconomy example in Regression and Other Stories.
}
\keyword{datasets}
