% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{incentives}
\alias{incentives}
\title{incentives}
\format{
data.frame: 62 obs. of 5 variables
}
\source{
\url{https://github.com/avehtari/ROS-Examples/tree/master/Incentives/data}
}
\usage{
incentives
}
\description{
Incentives data set for Incentives example in Regression and Other Stories.
}
\keyword{datasets}
