% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{kidiq}
\alias{kidiq}
\title{kidiq}
\format{
data.frame: 434 obs. of 5 variables
}
\source{
\url{https://github.com/avehtari/ROS-Examples/tree/master/KidIQ/data}
}
\usage{
kidiq
}
\description{
kidiq data set for KidIQ example in Regression and Other Stories.
}
\keyword{datasets}
