% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mesquite}
\alias{mesquite}
\title{mesquite}
\format{
data.frame: 46 obs. of 8 variables
}
\source{
\url{https://github.com/avehtari/ROS-Examples/tree/master/Mesquite/data}
}
\usage{
mesquite
}
\description{
Mesquite data set for Mesquite example in Regression and Other Stories.
}
\keyword{datasets}
