% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{mile}
\alias{mile}
\title{mile}
\format{
data.frame: 32 obs. of 6 variables
}
\source{
\url{https://github.com/avehtari/ROS-Examples/tree/master/Mile/data}
}
\usage{
mile
}
\description{
Mile data set for Mile example in Regression and Other Stories.
}
\keyword{datasets}
