% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{naes04}
\alias{naes04}
\title{naes04}
\format{
data.frame: 81422 obs. of 7 variables
}
\source{
\url{https://github.com/avehtari/ROS-Examples/tree/master/Gay/data}
}
\usage{
naes04
}
\description{
naes04 data set for Gay example in Regression and Other Stories.
}
\keyword{datasets}
