% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{nes}
\alias{nes}
\title{nes}
\format{
data.frame: 34908 obs. of 70 variables
}
\source{
\url{https://github.com/avehtari/ROS-Examples/tree/master/NES/data}
}
\usage{
nes
}
\description{
NES data set for NES example in Regression and Other Stories.
}
\keyword{datasets}
