% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{newcomb}
\alias{newcomb}
\title{newcomb}
\format{
data.frame: 66 obs. of 1 variable
}
\source{
\url{https://github.com/avehtari/ROS-Examples/tree/master/Newcomb/data}
}
\usage{
newcomb
}
\description{
Newcomb data set for Newcomb example in Regression and Other Stories.
}
\keyword{datasets}
