% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{oscars}
\alias{oscars}
\title{oscars}
\format{
data.frame: 1631 obs. of 62 variables.
}
\source{
\url{https://github.com/avehtari/ROS-Examples/tree/master/AcademyAwards/}
}
\usage{
oscars
}
\description{
Oscars data set for AcademyAwards example in Regression and Other Stories.
}
\keyword{datasets}
