% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{poll}
\alias{poll}
\title{poll}
\format{
data.frame: 778 obs. of 2 variables
}
\source{
\url{https://github.com/avehtari/ROS-Examples/tree/master/Poststrat/data}
}
\usage{
poll
}
\description{
Poll data set for Poststrat example in Regression and Other Stories.
}
\keyword{datasets}
