% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pollution}
\alias{pollution}
\title{pollution}
\format{
data.frame: 50 obs. of 16 variables
}
\source{
\url{https://github.com/avehtari/ROS-Examples/tree/master/Pollution/data}
}
\usage{
pollution
}
\description{
Pollution data set for Pollution example in Regression and Other Stories.
}
\keyword{datasets}
