% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{psprimaire}
\alias{psprimaire}
\title{psprimaire}
\format{
data.frame: 9 obs. of 4 variables
}
\source{
\url{https://github.com/avehtari/ROS-Examples/tree/master/FrenchElection/data}
}
\usage{
psprimaire
}
\description{
psprimaire data set for FrenchElection example in Regression and Other Stories.
}
\keyword{datasets}
