% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pyth}
\alias{pyth}
\title{pyth}
\format{
data.frame: 60 obs. of 3 variables
}
\source{
\url{https://github.com/avehtari/ROS-Examples/tree/master/Pyth}
}
\usage{
pyth
}
\description{
Pyth data set for Pyth example in Regression and Other Stories.
}
\keyword{datasets}
