% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{risky}
\alias{risky}
\title{risky}
\format{
data.frame: 434 obs. of 6 variables
}
\source{
\url{https://github.com/avehtari/ROS-Examples/tree/master/RiskyBehavior/data}
}
\usage{
risky
}
\description{
Risky data set for RiskyBehavior example in Regression and Other Stories.
}
\keyword{datasets}
