% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{riverbay}
\alias{riverbay}
\title{riverbay}
\format{
data.frame: 27 obs. of 8 variables
}
\source{
\url{https://github.com/avehtari/ROS-Examples/tree/master/Coop/data}
}
\usage{
riverbay
}
\description{
Riverbay data set for Coop example in Regression and Other Stories.
}
\keyword{datasets}
