% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{roaches}
\alias{roaches}
\title{roaches}
\format{
data.frame: 262 obs. of 6 variables
}
\source{
\url{https://github.com/avehtari/ROS-Examples/tree/master/Roaches/data}
}
\usage{
roaches
}
\description{
Roaches data set for Roaches example in Regression and Other Stories.
}
\keyword{datasets}
