% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sesame}
\alias{sesame}
\title{sesame}
\format{
data.frame: 240 obs. of 32 variables
}
\source{
\url{https://github.com/avehtari/ROS-Examples/tree/master/Sesame/data}
}
\usage{
sesame
}
\description{
Sesame data set for Sesame example in Regression and Other Stories.
}
\keyword{datasets}
