% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{SIS}
\alias{SIS}
\title{SIS}
\format{
data.frame: 1501 obs. of 15 variables
}
\source{
\url{https://github.com/avehtari/ROS-Examples/tree/master/Imputation/data}
}
\usage{
SIS
}
\description{
SIS data set for Imputation example in Regression and Other Stories.
}
\keyword{datasets}
