% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{student}
\alias{student}
\title{student}
\format{
data.frame: 382 obs. of 32 variables
}
\source{
\url{https://github.com/avehtari/ROS-Examples/tree/master/Student/data}
}
\usage{
student
}
\description{
Student data set for Student example in Regression and Other Stories.
}
\keyword{datasets}
