% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{unemp}
\alias{unemp}
\title{unemp}
\format{
data.frame: 70 obs. of 2 variables
}
\source{
\url{https://github.com/avehtari/ROS-Examples/tree/master/Unemployment/data/}
}
\usage{
unemp
}
\description{
unemp data set for Unemployment example in Regression and Other Stories.
}
\keyword{datasets}
