% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wells}
\alias{wells}
\title{wells}
\format{
data.frame: 3020 obs. of 7 variables
}
\source{
\url{https://github.com/avehtari/ROS-Examples/tree/master/Arsenic/data/}
}
\usage{
wells
}
\description{
Wells data set for Asrenic example in Regression and Other Stories.
}
\keyword{datasets}
